Rename enum LearnMoreUrl to LearnMoreUrlType for better readability
